test(federated): #285 split connection-init tests into duckdb_conn_init_test.go
Pure move: duckdb_conn_test.go was 531 lines, over the 500-line standard.
The #245/#285 init tests (FirstConnection/AllConnections/Concurrent,
recordingExecer, TestMakeConnInit_FailedInstallSkipsLoad, and the shared
newS3ConfiguredDuckDBConfig fixture) now live in a focused file.
